move logical plan to expr crate
