+++ datafusion/core/src/optimizer/common_subexpr_eliminate.rs
@@ -19,12 +19,12 @@
 
 use crate::error::Result;
 use crate::execution::context::ExecutionProps;
-use crate::logical_plan::plan::{Filter, Projection, Window};
+use crate::logical_expr::logical_plan::{
+    Aggregate, Filter, LogicalPlan, Projection, Sort, Window,
+};
 use crate::logical_plan::{
-    col,
-    plan::{Aggregate, Sort},
-    DFField, DFSchema, Expr, ExprRewritable, ExprRewriter, ExprSchemable, ExprVisitable,
-    ExpressionVisitor, LogicalPlan, Recursion, RewriteRecursion,
+    col, DFField, DFSchema, Expr, ExprRewritable, ExprRewriter, ExprSchemable,
+    ExprVisitable, ExpressionVisitor, Recursion, RewriteRecursion,
 };
 use crate::optimizer::optimizer::OptimizerRule;
 use crate::optimizer::utils;
