+++ datafusion/core/src/optimizer/eliminate_limit.rs
@@ -18,7 +18,7 @@
 //! Optimizer rule to replace `LIMIT 0` on a plan with an empty relation.
 //! This saves time in planning and executing the query.
 use crate::error::Result;
-use crate::logical_plan::{EmptyRelation, Limit, LogicalPlan};
+use crate::logical_expr::logical_plan::{EmptyRelation, Limit, LogicalPlan};
 use crate::optimizer::optimizer::OptimizerRule;
 
 use super::utils;
