+++ datafusion/core/src/physical_plan/hash_join.rs
@@ -63,7 +63,7 @@ use super::{
 };
 use super::{hash_utils::create_hashes, Statistics};
 use crate::error::{DataFusionError, Result};
-use crate::logical_plan::JoinType;
+use crate::logical_expr::logical_plan::JoinType;
 
 use super::{
     DisplayFormatType, ExecutionPlan, Partitioning, RecordBatchStream,
